+++ index.js
@@ -1,7 +1,3 @@
-/**
- * @format
- */
-
 import {AppRegistry} from 'react-native';
 import 'react-native-gesture-handler';
 import App from './App';
